wip - the big changes
