+++ help.py
@@ -12,7 +12,7 @@ TOOLS:
   ./help.py [[this file]]
 
   === Generate results and plot ===
-  ./tokenspice.py [[Run TokenSPICE simulation. Input netlist. Output rundir w/ csv.]]
+  ./tsp [[Run TokenSPICE simulation. Input netlist. Output rundir w/ csv.]]
   ./plot_1.py [[Plot results from 1 run. Input csv. Output pngs.]]
   ./showstats.py [[Show performance statistics for a run]]
   Example flow: see bottom
@@ -28,7 +28,7 @@ TOOLS:
   -Logging: ./logging.conf
 
   == Example flow ==
-  rm -rf outdir_csv; ./tokenspice.py mynetlist1.py outdir_csv 1>out.txt 2>&1 &
+  rm -rf outdir_csv; ./tsp mynetlist1.py outdir_csv 1>out.txt 2>&1 &
   tail -f out.txt
   rm -rf outdir_png; ./plot_1.py outdir_csv outdir_png
   eog outdir_png
